Add an `Edition::Searchable` concern
This removes the `Searchable` include and config from the base `Edition`
model and moves it to a concern. We can then include it in everything
that inherits the base `Edition` model. This will then allow us to
inherit `Edition`s without having to override search functionality if we
don’t need it.
