+++ app/models/case_study.rb
@@ -1,4 +1,6 @@
 class CaseStudy < Edition
+  include Edition::Searchable
+
   include Edition::Images
   include Edition::FactCheckable
   include Edition::CustomLeadImage
