+++ test/support/searchable_edition.rb
@@ -1,0 +1,5 @@
+require_relative "./generic_edition"
+
+class SearchableEdition < GenericEdition
+  include Edition::Searchable
+end
